fix Tls allocation bug
The Tls struct was allocated in the same block as the first task to run
in the new thread, meaning when that initial task exits and gets freed,
it takes Tls with it.
To fix this but not incur extra malloc overhead, the first new task
created in the thread, after the original one, copies the Tls struct
into a newly malloc'd buffer which gets freed when all tasks have
exited.